Move guess_table tests to the main test file
